move helper functions to scene-utils
